==> 1_experiment(train).ipynb <==
import pandas as pd 
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler,LabelEncoder
import pickle

#load the dataset
data=pd.read_csv("Churn_Modelling.csv")
data.head()

#preprocess the data
#drop irrelavant columns
data=data.drop(['RowNumber','CustomerId','Surname'],axis=1)  #axis=1 means we are dropping column wise
data

#encode categorical variable
label_encoder_gender=LabelEncoder()
data['Gender']=label_encoder_gender.fit_transform(data['Gender'])
data

#oneHot encode 'Geography'--> classified more than 2 class so applied one hotencoder
from sklearn.preprocessing import OneHotEncoder
onehot_encoder_geo=OneHotEncoder()
geo_encoder=onehot_encoder_geo.fit_transform(data[['Geography']])
geo_encoder


geo_encoder.toarray()

onehot_encoder_geo.get_feature_names_out(['Geography'])  # we will get the output of Geography into numpy

geo_encoded_df=pd.DataFrame(geo_encoder.toarray(),columns=onehot_encoder_geo.get_feature_names_out(['Geography']))
geo_encoded_df

#combine one hot encoder columns with the original data
data=pd.concat([data.drop('Geography',axis=1),geo_encoded_df],axis=1)  ##axis=1 means column wise
data.head()

#save the encoders and scaler in vs code as file
with open('label_encoder_gender.pkl','wb') as file:
    pickle.dump(label_encoder_gender,file)

with open('onehot_encoder_geo.pkl','wb') as file:
    pickle.dump(onehot_encoder_geo,file)


#divide the dataset into independent and dependent features
X=data.drop('Exited',axis=1)
y=data['Exited']

#split the data into training an dtesting sets
X_train,X_test,y_train,y_test=train_test_split(X,y,test_size=0.2,random_state=42) # Data is split the same way every time -->Results stay same

##scale these features
scaler=StandardScaler()
X_train=scaler.fit_transform(X_train)
X_test=scaler.transform(X_test)


X_train

#save scalar data as file in vs code   ##for preprocessing data we dont use model.save for saving
with open('scaler.pkl','wb') as file:  #wb-write binary
    pickle.dump(scaler,file)

data

#ANN CLASSIFICATION

import tensorflow as tf
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import Dense
from tensorflow.keras.callbacks import EarlyStopping,TensorBoard
import datetime


(X_train.shape[1],)  #(12,)means -->it is single dimension as it has 12 input

#build our ANN Model
model=Sequential([  #hidden layer 1 is connected with input layer
    Dense(64,activation='relu',input_shape=(X_train.shape[1],)), ## we are using 64 neurons apply relu as activation function and input size initailly we are giviing as x_train.shape (jitne v train data honge wo ho jyega input ka shape)
    Dense(32,activation='relu'), #hidden layer2
    Dense(1,activation='sigmoid') ##output layer
]
)


model.summary()   ##it will give total number of parameter(input*hiddenlayer+bias) in each layer  and at last total parameters it will give

import tensorflow
opt=tensorflow.keras.optimizers.Adam(learning_rate=0.01)   ##optimizers is for reduce the loss
loss=tensorflow.keras.losses.BinaryCrossentropy()  #loss function-how much error exists
loss

#compile the model(we are using adam optimizer here)
model.compile(optimizer="adam",loss="binary_crossentropy",metrics=['accuracy'])

##setup the tensorboard
from tensorflow.keras.callbacks import EarlyStopping,TensorBoard

log_dir="logs/fit"+datetime.datetime.now().strftime("%Y%m%d-%H%M%S")  ## Location where training logs are stored
tensorflow_callback=TensorBoard(log_dir=log_dir,histogram_freq=1)

##setup Early stopping
#  -->we use early stopping becasue let suppose our loss value is decreasing but after some time its stop decreasing so then we no need to run our model(early stopping will check if loss is not decreasing then we can stop the training there itself ) 
early_stopping_callback=EarlyStopping(monitor='val_loss',patience=5,restore_best_weights=True) ##patience will check the loss value still not decreasing upto 5 epoch then stop the training


##train the  model
history=model.fit(
    X_train,y_train,validation_data=(X_test,y_test),epochs=100,
    callbacks=[tensorflow_callback,early_stopping_callback]
)

model.save('model.h5')  ##we will use model.save for neural network save data

#load TensorBoard extension
%load_ext tensorboard

%reload_ext tensorboard

%tensorboard --logdir logs --port 6007

%tensorboard --logdir logs/fit20260205-163849
