==> 2_predection(test).ipynb <==
import numpy as mp
import pandas as pd
import tensorflow as tf
import pickle
from tensorflow.keras.models import load_model

##pickle converts a Python object into a file and later recreates the same object from that file
##Training an ML/ANN model--> Takes time ,Uses CPU/GPU So instead of training again and again, we save the trained model in pickel text file format later we convert pickle tp python file

##for loading pickle file (we have to load all the models )
model=load_model('model.h5')

##load the encoder and scalar 
with open('onehot_encoder_geo.pkl','rb') as file:  #rb-read binary
    onehot_encoder_geo=pickle.load(file)

with open('label_encoder_gender.pkl','rb') as file:
    label_encoder_gender=pickle.load(file)

with open('scaler.pkl','rb') as file:
    scaler=pickle.load(file)

input_data={
    'CreditScore':600,
    'Geography':'France',
    'Gender':'Male',
    'Age':40,
    'Tenure':3,
    'Balance':60000,
    'NumOfProducts':2,
    'HasCrCard':1,
    'IsActiveMember':1,
    'EstimatedSalary':5000
}

#one-hot encode 'Geography'
geo_encoded=onehot_encoder_geo.transform([[input_data['Geography']]]).toarray()
geo_encoded_df=pd.DataFrame(geo_encoded, columns=onehot_encoder_geo.get_feature_names_out(['Geography']))
geo_encoded_df


#the input data are in key-value pairs so we have to convert it into the dataframes
input_df=pd.DataFrame([input_data])
input_df

#Encode categorical variables(Gender into 0-->female and 1's-->Male)
input_df['Gender']=label_encoder_gender.transform(input_df['Gender'])
input_df

#combine one-hot encoded column with input data  (Germany,spain and france will be added after this concatenation with input_df in original data)
input_df=pd.concat([input_df.drop("Geography",axis=1),geo_encoded_df],axis=1)
input_df

#scaling the input data-->data will converted into array
input_scaled=scaler.transform(input_df)
input_scaled


##predection(will customer will leave the bank or not)
prediction=model.predict(input_scaled)
prediction


prediction_proba=prediction[0][0]  #array[0][0]=0.031146...

prediction_proba

if prediction_proba>0.5:
    print("The customer is likely to leave the bank")
else:
    print("the customer is not likely to leave the bank")
